+++ modules/communities/api/endpoints/getCommunityRequestList.ts
@@ -1,6 +1,6 @@
 import { type CoreEndpointBuilder } from 'appdeptus/api'
 import { type CommunityRequest } from 'appdeptus/models'
-import { supabase } from 'appdeptus/utils'
+import { mapNullToUndefined, supabase } from 'appdeptus/utils'
 import { Table } from 'appdeptus/utils/supabase'
 import { communityRequestListSchema } from '../schemas'
 import { type CommunitiesApiTags } from '../tags'
@@ -28,7 +28,9 @@ const getCommunityRequestList = (
           return { error: JSON.stringify(error) }
         }
 
-        const requests = await communityRequestListSchema.parseAsync(data)
+        const requests = await communityRequestListSchema.parseAsync(
+          mapNullToUndefined(data)
+        )
 
         return { data: requests }
       } catch (error) {