+++ CHANGELOG.md
@@ -13,6 +13,7 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 
 - Nickname and community forms to support special chars 🔧
 - Save button in army screen showing improperly 🫥
+- Community requests failing on users with no avatar set 🤳🏼
 
 ### Deprecated